personal deatails component complited
